Enhancements to front ends
