+++ TgaBuilderAvaloniaUi/View/BrickTransitionWindow.axaml
@@ -203,9 +203,9 @@
         <Grid Grid.Row="1">
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="Auto"/>
-                <ColumnDefinition Width="Auto"/>
-                <ColumnDefinition Width="Auto"/>
-                <ColumnDefinition Width="Auto"/>
+                <ColumnDefinition Width="100"/>
+                <ColumnDefinition Width="100"/>
+                <ColumnDefinition Width="100"/>
             </Grid.ColumnDefinitions>
             <Grid.RowDefinitions>
                 <RowDefinition Height="Auto"/>
@@ -223,12 +223,12 @@
                     VerticalAlignment="Center"
                     Text="Pivot"
                     FontStyle="Italic"/>
-                <ToggleSwitch
-                    Margin="5"
+                <CheckBox
+                    Margin="7"
                     VerticalAlignment="Center"
                     HorizontalAlignment="Left"
                     IsChecked="{Binding ReversePivot}"
-                    Content="Reverse"/>
+                    Content="Rev"/>
             </StackPanel>
 
             <Slider
