+++ TgaBuilderLib/Transitions/TransitionHelper.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 using System.Text;
 
 namespace TgaBuilderLib.Transitions
@@ -39,7 +40,44 @@ namespace TgaBuilderLib.Transitions
             MarkerRadius = 3; 
             ExpectedRegionCount = 0;
             ReversePivot = false;
-
         }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        // Computes directional distances to texture domains for the selected transition mode.
+        private (float distToT1, float distToT2) ComputeTopologicy(TransitionMode mode, float nx, float ny)
+            => mode switch
+            {
+                TransitionMode.Top => (
+                    Math.Min(nx, Math.Min(1.0f - nx, 1.0f - ny)),
+                    ny
+                ),
+
+                TransitionMode.Bottom => (
+                    Math.Min(nx, Math.Min(1.0f - nx, ny)),
+                    1.0f - ny
+                ),
+
+                TransitionMode.Left => (
+                    Math.Min(ny, Math.Min(1.0f - ny, 1.0f - nx)),
+                    nx
+                ),
+
+                TransitionMode.Right => (
+                    Math.Min(ny, Math.Min(1.0f - ny, nx)),
+                    1.0f - nx
+                ),
+
+                TransitionMode.DiagonalTopLeft => (
+                    Math.Min(1.0f - nx, 1.0f - ny),
+                    Math.Min(nx, ny)
+                ),
+
+                TransitionMode.DiagonalTopRight => (
+                    Math.Min(nx, 1.0f - ny),
+                    Math.Min(1.0f - nx, ny)
+                ),
+
+                _ => (0f, 0f)
+            };
     }
 }
